feat: Add authorization to deleteDepartment and deleteEmployee APIs
This commit adds authorization to the `deleteDepartment` and `deleteEmployee` APIs in the `DepartmentController` and `EmployeeController` respectively. The `deleteDepartment` and `deleteEmployee` methods now check the role of the user making the request and only allow deletion if the user has the HR or ADMIN role. If the user does not have the required role, a `HttpException` with a 403 status code and the message "Forbidden" is thrown. This ensures that only authorized users can delete departments and employees.