File upload (#13)
* chore: Add file upload functionality

This commit adds the functionality to upload files using the multer library. It includes the necessary changes to the package.json file to add the multer dependency, as well as the addition of the fileUpload.ts utility file. The fileUpload.ts file contains the logic for handling file uploads, including memory buffer storage, and validating the file type and mimetype. Additionally, also runs a transaction for uploading books to db from xlsx data.

* feat: Add file upload functionality for bulk book upload